Factor out various data structures from parallel_envs to their own modules
