+++ gridengine/src/gridengine/parallel_environments.py
@@ -1,56 +1,23 @@
-# for backwards compatability, this is a global var instead of an argument
-# to sge_job_handler.
-import re
-import typing
-from abc import ABC, abstractmethod
-from io import StringIO
-from shutil import which
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict
 
 from hpc.autoscale import hpclogging as logging
-from hpc.autoscale import hpctypes as ht
-from hpc.autoscale.node.constraints import Constraint
-from hpc.autoscale.util import partition_single
 
-from gridengine.util import check_output
-
-if typing.TYPE_CHECKING:
-    from gridengine.environment import GridEngineEnvironment
-
-QCONF_PATH = which("qconf") or ""
-
-
-_PE_CACHE: Dict[str, "ParallelEnvironment"] = {}
-_QUEUE_CACHE: List["GridEngineQueue"] = []
-_COMPLEX_CACHE: Dict[str, "Complex"] = {}
-
-
-def set_pe(pe_name: str, pe: Union[Dict[str, str], "ParallelEnvironment"]) -> None:
-    global _PE_CACHE
-    if _PE_CACHE is None:
-        _PE_CACHE = {}
-
-    if not isinstance(pe, ParallelEnvironment):
-        pe = ParallelEnvironment(pe)
-
-    _PE_CACHE[pe_name] = pe
+from gridengine.allocation_rules import AllocationRule
+from gridengine.qbin import QBin
+from gridengine.util import parse_ge_config
 
 
 def read_parallel_environments(
-    autoscale_config: Dict,
+    autoscale_config: Dict, qbin: QBin,
 ) -> Dict[str, "ParallelEnvironment"]:
-    global _PE_CACHE
-    if _PE_CACHE:
-        return _PE_CACHE
-
     parallel_envs = {}
     pe_config = autoscale_config.get("gridengine", {}).get("pes", {})
-    pe_names = check_output([QCONF_PATH, "-spl"]).decode().splitlines()
+    pe_names = qbin.qconf(["-spl"]).splitlines()
 
     for pe_name in pe_names:
         pe_name = pe_name.strip()
-        lines = check_output([QCONF_PATH, "-sp", pe_name]).decode().splitlines(False)
-        pe = _parse_ge_config(lines)
+        lines = qbin.qconf(["-sp", pe_name]).splitlines(False)
+        pe = parse_ge_config(lines)
 
         req_key = "requires_placement_groups"
 
@@ -64,322 +31,7 @@ def read_parallel_environments(
 
         parallel_envs[pe_name] = ParallelEnvironment(pe)
 
-    _PE_CACHE = parallel_envs
-    return _PE_CACHE
-
-
-def read_queue_configs(autoscale_config: Dict, ge_env: "GridEngineEnvironment") -> None:
-    qnames = check_output([QCONF_PATH, "-sql"]).decode().split()
-
-    logging.debug("Found %d queues: %s", len(qnames), " ".join(qnames))
-    autoscale_queues_config = autoscale_config.get("gridengine", {}).get("queues", {})
-
-    for qname in qnames:
-        lines = check_output([QCONF_PATH, "-sq", qname]).decode().splitlines()
-        queue_config = _parse_ge_config(lines)
-        queue_constraints = autoscale_queues_config.get(queue_config["qname"], {}).get(
-            "constraints", []
-        )
-        ge_env.queues[qname] = GridEngineQueue(queue_config, ge_env, queue_constraints)
-
-
-def set_complexes(autoscale_config: Dict, complex_lines: List[str]) -> None:
-    global _COMPLEX_CACHE
-    _COMPLEX_CACHE = _parse_complexes(autoscale_config, complex_lines)
-
-
-def read_complexes(autoscale_config: Dict) -> Dict[str, "Complex"]:
-    global _COMPLEX_CACHE
-    if _COMPLEX_CACHE:
-        return _COMPLEX_CACHE
-
-    complex_lines = check_output([QCONF_PATH, "-sc"]).decode().splitlines()
-    _COMPLEX_CACHE = _parse_complexes(autoscale_config, complex_lines)
-    return _COMPLEX_CACHE
-
-
-def _parse_complexes(
-    autoscale_config: Dict, complex_lines: List[str]
-) -> Dict[str, "Complex"]:
-    relevant_complexes = None
-    if autoscale_config:
-        relevant_complexes = autoscale_config.get("gridengine", {}).get(
-            "relevant_complexes"
-        )
-        if relevant_complexes:
-            logging.info(
-                "Restricting complexes for autoscaling to %s", relevant_complexes
-            )
-
-    complexes: List[Complex] = []
-    headers = complex_lines[0].lower().replace("#", "").split()
-
-    required = set(["name", "type", "consumable"])
-    missing = required - set(headers)
-    if missing:
-        logging.error(
-            "Could not parse complex file as it is missing expected columns: %s."
-            + " Autoscale likely will not work.",
-            list(missing),
-        )
-        return {}
-
-    for n, line in enumerate(complex_lines[1:]):
-        if line.startswith("#"):
-            continue
-        toks = line.split()
-        if len(toks) != len(headers):
-            logging.warning(
-                "Could not parse complex at line {} - ignoring: '{}'".format(n, line)
-            )
-            continue
-        c = dict(zip(headers, toks))
-        try:
-            if relevant_complexes and c["name"] not in relevant_complexes:
-                logging.trace(
-                    "Ignoring complex %s because it was not defined in gridengine.relevant_complexes",
-                    c["name"],
-                )
-                continue
-
-            complex = Complex(
-                name=c["name"],
-                shortcut=c.get("shortcut", c["name"]),
-                complex_type=c["type"],
-                relop=c.get("relop", "=="),
-                requestable=c.get("requestable", "YES").lower() == "yes",
-                consumable=c.get("consumable", "YES").lower() == "yes",
-                default=c.get("default"),
-                urgency=int(c.get("urgency", 0)),
-            )
-
-            complexes.append(complex)
-
-        except Exception:
-            logging.exception("Could not parse complex %s - %s", line, c)
-
-    # TODO test RDH
-    ret = partition_single(complexes, lambda x: x.name)
-    shortcut_dict = partition_single(complexes, lambda x: x.shortcut)
-    ret.update(shortcut_dict)
-    return ret
-
-
-def _flatten_lines(lines: List[str]) -> List[str]:
-    ret = []
-    n = 0
-    while n < len(lines):
-        line = lines[n]
-        while line.endswith("\\") and n < len(lines) - 1:
-            n += 1
-            next_line = lines[n]
-            line = line[:-1] + " " + next_line.strip()
-        ret.append(line)
-        n += 1
-    return ret
-
-
-def _parse_ge_config(lines: List[str]) -> Dict[str, str]:
-    config = {}
-
-    for line in _flatten_lines(lines):
-        toks = re.split(r"[ \t]+", line, 1)
-        if len(toks) != 2:
-            continue
-        config[toks[0].strip()] = toks[1].strip()
-
-    return config
-
-
-def new_gridenging_queue(
-    qname: str,
-    hostlist: List[str],
-    pe_list: List[str],
-    constraints: List[Constraint],
-    slots_expr: str,
-    ge_env: "GridEngineEnvironment",
-) -> "GridEngineQueue":
-    assert isinstance(hostlist, list)
-    queue_config = {
-        "qname": qname,
-        "hostlist": " ".join(hostlist),
-        "pe_list": " ".join(pe_list),
-        "slots": slots_expr,
-    }
-    return GridEngineQueue(queue_config, ge_env, constraints)
-
-
-class GridEngineQueue:
-    def __init__(
-        self,
-        queue_config: Dict,
-        ge_env: "GridEngineEnvironment",
-        constraints: List[Constraint],
-    ) -> None:
-        self.queue_config = queue_config
-        self.ge_env = ge_env
-        if isinstance(constraints, dict):
-            constraints = [constraints]
-
-        if not isinstance(constraints, list):
-            logging.error(
-                "Ignoring invalid constraints for queue %s! %s", self.qname, constraints
-            )
-            constraints = [{"node.nodearray": self.qname}]
-
-        self.constraints = constraints
-        self.__hostlist = re.split(",| +", self.queue_config["hostlist"])
-        self.__pe_to_hostgroups: Dict[str, List[str]] = {}
-        self._pe_keys_cache: Dict[str, List[str]] = {}
-        self.__parallel_environments: Dict[str, "ParallelEnvironment"] = {}
-        self.__slots = {}
-
-        slot_toks = re.split(",| +", self.queue_config.get("slots", ""))
-        for tok in slot_toks:
-            if "[" not in tok:
-                continue
-            tok = tok.replace("[", "").replace("]", "")
-            fqdn, slot_str = tok.split("=")
-            slot = int(slot_str.strip())
-            hostname = fqdn.strip().lower().split(".")[0]
-            self.__slots[ht.Hostname(hostname)] = slot
-
-        for tok in tokenize_pe_list(queue_config["pe_list"]):
-            tok = tok.strip()
-
-            if tok == "NONE":
-                continue
-
-            pe_name_expr: str
-            hostgroups: List[str]
-
-            if "@" in tok and "=" in tok:
-                tok = tok.lstrip("[").rstrip("]")
-                hostgroup, pe_name_expr = tok.split("=", 1)
-
-                hostgroups = [hostgroup]
-            else:
-                pe_name_expr = tok
-                hostgroups = [x for x in self.__hostlist if x.startswith("@")]
-
-            pe_names = re.split(",| +", pe_name_expr)
-
-            for pe_name in pe_names:
-                if pe_name not in ge_env.pes:
-                    logging.warning(
-                        "Parallel environment %s was not found - %s. Skipping",
-                        pe_name,
-                        list(ge_env.pes.keys()),
-                    )
-                    continue
-                assert pe_name not in self.__pe_to_hostgroups
-                self.__pe_to_hostgroups[pe_name] = hostgroups
-                self.__parallel_environments[pe_name] = ge_env.pes[pe_name]
-
-        if queue_config["pe_list"]:
-            assert self.__parallel_environments, queue_config["pe_list"]
-
-        all_host_groups = set(self.hostlist)
-        for pe in self.__parallel_environments.values():
-            if pe.requires_placement_groups:
-                all_host_groups = all_host_groups - set(
-                    self.get_hostgroups_for_pe(pe.name)
-                )
-        self.__ht_hostgroups = list(all_host_groups)
-
-    @property
-    def qname(self) -> str:
-        return self.queue_config["qname"]
-
-    @property
-    def hostlist(self) -> List[str]:
-        return self.__hostlist
-
-    @property
-    def hostlist_groups(self) -> List[str]:
-        return [x for x in self.hostlist if x.startswith("@")]
-
-    @property
-    def slots(self) -> Dict[ht.Hostname, int]:
-        return self.__slots
-
-    @property
-    def complexes(self) -> Dict[str, "Complex"]:
-        # TODO RDH should not need this here
-        return self.ge_env.complexes
-
-    def has_pe(self, pe_name: str) -> bool:
-        return bool(self._pe_keys(pe_name))
-
-    def _pe_keys(self, pe_name: str) -> List[str]:
-        if "*" not in pe_name:
-            return [pe_name]
-
-        if pe_name not in self._pe_keys_cache:
-            pe_name_pat = pe_name.replace("*", ".*")
-            pe_name_re = re.compile(pe_name_pat)
-
-            ret = []
-
-            for key in self.__parallel_environments:
-
-                if pe_name_re.match(key):
-                    ret.append(key)
-
-            self._pe_keys_cache[pe_name] = ret
-
-        return self._pe_keys_cache[pe_name]
-
-    def get_pes(self, pe_name: Optional[str] = None) -> List["ParallelEnvironment"]:
-        if pe_name is None:
-            return list(self.__parallel_environments.values())
-
-        if not self.has_pe(pe_name):
-            raise RuntimeError(
-                "Queue {} does not support parallel_environment {}".format(
-                    self.qname, pe_name
-                )
-            )
-        return [self.__parallel_environments[x] for x in self._pe_keys(pe_name)]
-
-    def get_hostgroups_for_ht(self) -> List[str]:
-        # TODO expensive
-        return self.__ht_hostgroups
-
-    def get_hostgroups_for_pe(self, pe_name: str) -> List[str]:
-        if not self.has_pe(pe_name):
-            raise RuntimeError(
-                "Queue {} does not support parallel_environment {}".format(
-                    self.qname, pe_name
-                )
-            )
-
-        return self.__pe_to_hostgroups[pe_name]
-
-    def get_placement_group(self, pe_name: str) -> Optional[str]:
-        if pe_name not in self.__parallel_environments:
-            raise RuntimeError(
-                "Could not find pe {} - {}".format(
-                    pe_name, list(self.__parallel_environments.keys())
-                )
-            )
-        pe = self.__parallel_environments[pe_name]
-
-        if not pe.requires_placement_groups:
-            return None
-
-        first = self.get_hostgroups_for_pe(pe.name)[0]
-        placement_group = ht.PlacementGroup(first.replace("@", ""))
-        placement_group = re.sub("[^a-zA-z0-9-_]", "_", placement_group)
-        return placement_group
-
-    def __str__(self) -> str:
-        return "GridEnineQueue(qname={}, pes={}, hostlist={})".format(
-            self.qname, list(self.__parallel_environments.keys()), self.hostlist
-        )
-
-    def __repr__(self) -> str:
-        return str(self)
+    return parallel_envs
 
 
 def new_parallel_environment(
@@ -431,212 +83,3 @@ class ParallelEnvironment:
             self.allocation_rule,
             self.requires_placement_groups,
         )
-
-
-class AllocationRule(ABC):
-    def __init__(self, name: str) -> None:
-        self.__name = name
-
-    @property
-    @abstractmethod
-    def requires_placement_groups(self) -> bool:
-        ...
-
-    @property
-    def name(self) -> str:
-        return self.__name
-
-    @property
-    def is_fixed(self) -> bool:
-        return False
-
-    @staticmethod
-    def value_of(allocation_rule: str) -> "AllocationRule":
-        if allocation_rule == "$fill_up":
-            return FillUp()
-        elif allocation_rule == "$round_robin":
-            return RoundRobin()
-        elif allocation_rule == "$pe_slots":
-            return PESlots()
-        elif allocation_rule.isdigit():
-            return FixedProcesses(allocation_rule)
-        else:
-            raise RuntimeError("Unknown allocation rule {}".format(allocation_rule))
-
-    def __repr__(self) -> str:
-        cname = self.__class__.__name__
-        return "{}({})".format(cname, self.name)
-
-
-class FillUp(AllocationRule):
-    def __init__(self) -> None:
-        super().__init__("$fill_up")
-
-    @property
-    def requires_placement_groups(self) -> bool:
-        return True
-
-
-class RoundRobin(AllocationRule):
-    def __init__(self) -> None:
-        super().__init__("$round_robin")
-
-    @property
-    def requires_placement_groups(self) -> bool:
-        return True
-
-
-class PESlots(AllocationRule):
-    def __init__(self) -> None:
-        super().__init__("$pe_slots")
-
-    @property
-    def requires_placement_groups(self) -> bool:
-        return False
-
-
-class FixedProcesses(AllocationRule):
-    def __init__(self, name: str) -> None:
-        super().__init__(name)
-        self.__fixed_processes = int(name)
-
-    @property
-    def requires_placement_groups(self) -> bool:
-        return True
-
-    @property
-    def is_fixed(self) -> bool:
-        return True
-
-    @property
-    def fixed_processes(self) -> int:
-        return self.__fixed_processes
-
-
-class Complex:
-    def __init__(
-        self,
-        name: str,
-        shortcut: str,
-        complex_type: str,
-        relop: str,
-        requestable: bool,
-        consumable: bool,
-        default: Optional[str],
-        urgency: int,
-    ) -> None:
-        self.name = name
-        self.shortcut = shortcut
-        self.complex_type = complex_type.upper()
-        self.relop = relop
-        self.requestable = requestable
-        self.consumable = consumable
-        self.urgency = urgency
-        self.__logged_type_warning = False
-        self.__logged_parse_warning = False
-
-        self.default = self.parse(default or "NONE")
-
-    def parse(self, value: str) -> Optional[Any]:
-        try:
-            if value.upper() == "NONE":
-                return None
-            if value.lower() == "infinity":
-                return float("inf")
-
-            if self.complex_type in ["INT", "RSMAP"]:
-                return int(value)
-
-            elif self.complex_type == "BOOL":
-                try:
-                    return bool(float(value))
-                except ValueError:
-                    if value.lower() in ["true", "false"]:
-                        return value.lower() == "true"
-                    else:
-                        logging.warning(
-                            "Could not parse '%s' for complex type %s - treating as string.",
-                            value,
-                            self.complex_type,
-                        )
-                    return value
-            elif self.complex_type == "DOUBLE":
-                return float(value)
-
-            elif self.complex_type in ["RESTRING", "TIME", "STRING", "HOST"]:
-                return value
-
-            elif self.complex_type == "CSTRING":
-                # TODO test
-                return value.lower()  # case insensitve - we will just always lc
-
-            elif self.complex_type == "MEMORY":
-                size = value[-1]
-                if size.isdigit():
-                    mem = ht.Memory(float(value), "b")
-                else:
-                    mem = ht.Memory(float(value[:-1]), size)
-                return mem.convert_to("g")
-            else:
-                if not self.__logged_type_warning:
-                    logging.warning(
-                        "Unknown complex type %s - treating as string.",
-                        self.complex_type,
-                    )
-                    self.__logged_type_warning = True
-                return value
-        except Exception:
-            if not self.__logged_parse_warning:
-                logging.warning(
-                    "Could not parse complex %s with value '%s'. Treating as string",
-                    self,
-                    value,
-                )
-                self.__logged_parse_warning = True
-            return value
-
-    def __str__(self) -> str:
-        return "Complex(name={}, shortcut={}, type={}, default={})".format(
-            self.name, self.shortcut, self.complex_type, self.default
-        )
-
-    def __repr__(self) -> str:
-        return "Complex(name={}, shortcut={}, type={}, relop='{}', requestable={}, consumable={}, default={}, urgency={})".format(
-            self.name,
-            self.shortcut,
-            self.complex_type,
-            self.relop,
-            self.requestable,
-            self.consumable,
-            self.default,
-            self.urgency,
-        )
-
-
-def tokenize_pe_list(expr: str) -> List[str]:
-    buf = StringIO()
-
-    in_left_bracket = False
-
-    toks = []
-
-    for c in expr:
-        if c.isalnum():
-            buf.write(c)
-            continue
-        if c == "[":
-            in_left_bracket = True
-            continue
-        if c == "]":
-            in_left_bracket = False
-
-        if c in [" ", ","] and not in_left_bracket:
-            toks.append(buf.getvalue())
-            buf = StringIO()
-        else:
-            buf.write(c)
-
-    if buf.getvalue():
-        toks.append(buf.getvalue())
-        buf = StringIO()
-    return toks
