+++ gridengine/test/gridengine_test/autoscaler_test.py
@@ -3,27 +3,23 @@ from typing import Dict, List, Optional
 
 from hpc.autoscale import hpclogging
 from hpc.autoscale.ccbindings.mock import MockClusterBinding
+from hpc.autoscale.hpctypes import Memory
 from hpc.autoscale.job.demandcalculator import DemandCalculator
 from hpc.autoscale.job.job import Job
 from hpc.autoscale.results import DefaultContextHandler, register_result_handler
 from hpc.autoscale.util import partition, partition_single
 
-from gridengine import autoscaler, parallel_environments
+from gridengine import autoscaler
+from gridengine.allocation_rules import FillUp, FixedProcesses, RoundRobin
+from gridengine.complex import Complex
 from gridengine.environment import GridEngineEnvironment
-from gridengine.parallel_environments import FillUp, FixedProcesses, RoundRobin
-from gridengine.parallel_environments import new_gridenging_queue as new_gequeue
 from gridengine.parallel_environments import new_parallel_environment as new_pe
+from gridengine.queue import new_gequeue
 from gridengine_test import mock_driver
 
-SLOTS_COMPLEX = parallel_environments.Complex(
-    "slots", "s", "INT", "<=", True, True, "1", 1000
-)
-MFREE_COMPLEX = parallel_environments.Complex(
-    "m_mem_free", "mfree", "MEMORY", "<=", True, True, "0", 0
-)
-EXCL_COMPLEX = parallel_environments.Complex(
-    "exclusive", "excl", "BOOL", "EXCL", True, True, "0", 1000
-)
+SLOTS_COMPLEX = Complex("slots", "s", "INT", "<=", True, True, "1", 1000)
+MFREE_COMPLEX = Complex("m_mem_free", "mfree", "MEMORY", "<=", True, True, "0", 0)
+EXCL_COMPLEX = Complex("exclusive", "excl", "BOOL", "EXCL", True, True, "0", 1000)
 CONTEXT = DefaultContextHandler("[default]")
 
 
@@ -84,6 +80,19 @@ def test_non_exclusive_htc_jobs() -> None:
     common_cluster_test(100 * ["-l slots=2 -q htc.q sleep.sh"], htc=10)
 
 
+def test_complex_shortcut_parsing() -> None:
+    ge_env = common_ge_env()
+    qsub1 = mock_driver.MockQsub(ge_env)
+    qsub1.qsub("-l m_mem_free=2g -l exclusive=true -q htc.q sleep.sh")
+    qsub2 = mock_driver.MockQsub(ge_env)
+    qsub2.qsub("-l mfree=2g -l excl=true -q  htc.q sleep.sh")
+    longform = qsub1.parse_jobs()
+    shortform = qsub2.parse_jobs()
+    assert longform
+    assert "mfree" not in str(longform)
+    assert longform[0].to_dict() == shortform[0].to_dict()
+
+
 def test_complex_shortcut() -> None:
     # make sure that if a user mixes the shortcut and long form
     # we still handle that.
@@ -95,84 +104,95 @@ def test_complex_shortcut() -> None:
             "-l mfree=2g      -q htc.q sleep.sh",
             "-l mfree=2g      -q htc.q sleep.sh",
             "-l mfree=2g      -q htc.q sleep.sh",
+            # "-l m_mem_free=2g -q htc.q sleep.sh",
+            # "-l m_mem_free=2g -q htc.q sleep.sh",
+            # "-l m_mem_free=2g -q htc.q sleep.sh",
         ],
         htc=2,
     )
+    eg = dcalc.node_mgr.example_node("westus", "Standard_F4")
+
     new_nodes = dcalc.get_demand().new_nodes
     by_name = partition_single(new_nodes, lambda n: n.name)
 
-    assert by_name["htc-1"].resources["m_mem_free"] == 8 * 1024
-    assert by_name["htc-1"].resources["mfree"] == 8 * 1024
-    assert by_name["htc-1"].available["m_mem_free"] == 0
-    assert by_name["htc-1"].available["mfree"] == 0
+    def m(expr: str) -> Memory:
+        return Memory.value_of(expr)
 
-    assert by_name["htc-2"].resources["m_mem_free"] == 8 * 1024
-    assert by_name["htc-2"].resources["mfree"] == 8 * 1024
-    assert by_name["htc-2"].available["m_mem_free"] == 6 * 1024
-    assert by_name["htc-2"].available["mfree"] == 6 * 1024
+    assert eg.memory == m("8g")
+
+    assert by_name["htc-1"].memory == m("8g")
+    assert by_name["htc-1"].resources["m_mem_free"] == m("8g")
+    assert by_name["htc-1"].resources["mfree"] == m("8g")
+    assert by_name["htc-1"].available["m_mem_free"] == m("0g")
+    assert by_name["htc-1"].available["mfree"] == m("0g")
+
+    assert by_name["htc-2"].resources["m_mem_free"] == m("8g")
+    assert by_name["htc-2"].resources["mfree"] == m("8g")
+    assert by_name["htc-2"].available["m_mem_free"] == m("4g")
+    assert by_name["htc-2"].available["mfree"] == m("4g")
 
 
-def test_round_robin() -> None:
-    # first test with RoundRobin (rr*)
+def test_fixed() -> None:
     # ask for more than the available count 100
-    common_cluster_test(["-l exclusive=1 -pe rr* 101  -q hpc.q sleep.sh"])
+    common_cluster_test(["-pe fp* 101  -q hpc.q sleep.sh"])
 
     # ask for more than the max vmss size
-    common_cluster_test(["-l exclusive=1 -pe rr* 7  -q hpc.q sleep.sh"])
+    common_cluster_test(["-pe fp* 14  -q hpc.q sleep.sh"])
 
     # ask for exactly the max vmss size
-    common_cluster_test(
-        ["-l exclusive -pe rr* 5  -q hpc.q sleep.sh"], {"hpc_q_rr2": 5}, hpc=5
-    )
+    common_cluster_test(["-pe fp* 10  -q hpc.q sleep.sh"], {"hpc_q_fp0": 5}, hpc=5)
 
     # same, except split across two jobs
     common_cluster_test(
         [
-            "-l exclusive=true -pe rr* 2  -q hpc.q sleep.sh",
-            "-l exclusive=true -pe rr* 3  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe fp* 4  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe fp* 6  -q hpc.q sleep.sh",
         ],
-        {"hpc_q_rr2": 5},
+        {"hpc_q_fp0": 5},
         hpc=5,
     )
 
     # let's allocate in two different pgs
     common_cluster_test(
         [
-            "-l exclusive=true -pe rr* 3  -q hpc.q sleep.sh",
-            "-l exclusive=true -pe rr* 3  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe fp* 6  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe fp* 6  -q hpc.q sleep.sh",
         ],
-        {"hpc_q_rr2": 3, "hpc_q_rr1": 3},
+        {"hpc_q_fp0": 3, "hpc_q_fp1": 3},
         hpc=6,
     )
 
     # let's allocate in all three pgs plus reach capacity.
     common_cluster_test(
-        ["-l exclusive=true -pe rr* 3 -q hpc.q sleep.sh"] * 4,
-        pg_counts={"hpc_q_rr2": 3, "hpc_q_rr1": 3, "hpc_q_rr0": 3},
+        ["-l exclusive=true -pe fp* 3 -q hpc.q sleep.sh"] * 4,
+        pg_counts={"hpc_q_fp2": 3, "hpc_q_fp1": 3, "hpc_q_fp0": 3},
         hpc=9,
     )
 
     # same as above, but let's use an array
     common_cluster_test(
-        ["-l exclusive=true -pe rr* 3 -t 1-4 -q hpc.q sleep.sh"],
-        pg_counts={"hpc_q_rr2": 3, "hpc_q_rr1": 3, "hpc_q_rr0": 3},
+        ["-l exclusive=true -pe fp* 3 -t 1-4 -q hpc.q sleep.sh"] * 4,
+        pg_counts={"hpc_q_fp2": 3, "hpc_q_fp1": 3, "hpc_q_fp0": 3},
         hpc=9,
     )
 
 
-def test_fill_up() -> None:
+def test_fill_up_and_round_robin() -> None:
     # With FillUp, GE will spread the processes across the machines as tightly as possible.
     # We are using an F4 here, so slots=4
     #
-    common_cluster_test(["-l exclusive=1 -pe fu* 40  -q hpc.q sleep.sh"], pg_counts={})
+    common_cluster_test(["-pe fu* 40  -q hpc.q sleep.sh"], pg_counts={})
+    common_cluster_test(["-l -pe rr* 40  -q hpc.q sleep.sh"], pg_counts={})
     # ask for more than the max vmss size
-    common_cluster_test(["-l exclusive=1 -pe fu* 28  -q hpc.q sleep.sh"], pg_counts={})
+    common_cluster_test(["-pe fu* 28  -q hpc.q sleep.sh"], pg_counts={})
+    common_cluster_test(["-pe rr* 28  -q hpc.q sleep.sh"], pg_counts={})
 
     # ask for exactly the max vmss size
     common_cluster_test(
-        ["-l exclusive=1 -pe fu* 20  -q hpc.q sleep.sh"],
-        pg_counts={"hpc_q_fu2": 5},
-        hpc=5,
+        ["-pe fu* 20  -q hpc.q sleep.sh"], pg_counts={"hpc_q_fu0": 5}, hpc=5,
+    )
+    common_cluster_test(
+        ["-pe rr* 20  -q hpc.q sleep.sh"], pg_counts={"hpc_q_rr0": 5}, hpc=5,
     )
 
     # same, except split across two jobs
@@ -181,7 +201,15 @@ def test_fill_up() -> None:
             "-l exclusive=true -pe fu* 8  -q hpc.q sleep.sh",
             "-l exclusive=true -pe fu* 12  -q hpc.q sleep.sh",
         ],
-        pg_counts={"hpc_q_fu2": 5},
+        pg_counts={"hpc_q_fu0": 5},
+        hpc=5,
+    )
+    common_cluster_test(
+        [
+            "-l exclusive=true -pe rr* 8  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe rr* 12  -q hpc.q sleep.sh",
+        ],
+        pg_counts={"hpc_q_rr0": 5},
         hpc=5,
     )
 
@@ -191,7 +219,15 @@ def test_fill_up() -> None:
             "-l exclusive=true -pe fu* 12  -q hpc.q sleep.sh",
             "-l exclusive=true -pe fu* 12  -q hpc.q sleep.sh",
         ],
-        pg_counts={"hpc_q_fu2": 3, "hpc_q_fu1": 3},
+        pg_counts={"hpc_q_fu0": 3, "hpc_q_fu1": 3},
+        hpc=6,
+    )
+    common_cluster_test(
+        [
+            "-l exclusive=true -pe rr* 12  -q hpc.q sleep.sh",
+            "-l exclusive=true -pe rr* 12  -q hpc.q sleep.sh",
+        ],
+        pg_counts={"hpc_q_rr0": 3, "hpc_q_rr1": 3},
         hpc=6,
     )
 
@@ -201,11 +237,21 @@ def test_fill_up() -> None:
         pg_counts={"hpc_q_fu2": 3, "hpc_q_fu1": 3, "hpc_q_fu0": 3},
         hpc=9,
     )
+    common_cluster_test(
+        ["-l exclusive=true -pe rr* 12 -q hpc.q sleep.sh"] * 4,
+        pg_counts={"hpc_q_rr2": 3, "hpc_q_rr1": 3, "hpc_q_rr0": 3},
+        hpc=9,
+    )
 
     # same as above, but let's use an array
     common_cluster_test(
         ["-l exclusive=true -pe fu* 12 -t 1-4 -q hpc.q sleep.sh"],
         pg_counts={"hpc_q_fu2": 3, "hpc_q_fu1": 3, "hpc_q_fu0": 3},
+        hpc=9,
+    )
+    common_cluster_test(
+        ["-l exclusive=true -pe rr* 12 -t 1-4 -q hpc.q sleep.sh"],
+        pg_counts={"hpc_q_rr2": 3, "hpc_q_rr1": 3, "hpc_q_rr0": 3},
         hpc=9,
     )
 
@@ -249,8 +295,8 @@ def mock_config(bindings: MockClusterBinding) -> Dict:
         "logging": {"config_file": logging_config},
         "default_resources": [
             {"name": "slots", "select": {}, "value": "node.vcpu_count"},
-            {"name": "mfree", "select": {}, "value": "node.resources.memgb"},
             {"name": "m_mem_free", "select": {}, "value": "node.resources.memgb"},
+            {"name": "mfree", "select": {}, "value": "node.resources.m_mem_free"},
         ],
         "nodearrays": {"hpc": {"placement_groups": pgs}},
     }
@@ -270,6 +316,12 @@ def common_cluster_test(
     by_array = partition(demand.new_nodes, lambda n: n.nodearray)
     by_pg = partition(demand.new_nodes, lambda n: n.placement_group)
     if set(by_pg.keys()) != set([None]):
+        if set(by_pg.keys()) != set(pg_counts.keys()):
+            assert False, "\n%s\n%s" % (
+                [(x, len(y)) for x, y in by_pg.items()],
+                pg_counts,
+            )
+        assert set(by_pg.keys()) == set(pg_counts.keys())
         assert not (bool(by_pg) ^ bool(pg_counts))
 
     if pg_counts:
@@ -277,8 +329,11 @@ def common_cluster_test(
             assert pg_name in by_pg
             assert (
                 len(by_pg[pg_name]) == count
-            ), "Expected pg {} to have {} nodes. Found {}".format(
-                pg_name, count, len(by_pg[pg_name])
+            ), "Expected pg {} to have {} nodes. Found {}. Full {}".format(
+                pg_name,
+                count,
+                len(by_pg[pg_name]),
+                [(x, len(y)) for x, y in by_pg.items()],
             )
 
         for pg_name in by_pg:
